+++ packages/react-native-reanimated/src/mutables.native.ts
@@ -1,6 +1,9 @@
 'use strict';
 
-import type { ShareableGuest, Synchronizable } from 'react-native-worklets';
+import type {
+  FixedSynchronizable,
+  ShareableGuest,
+} from 'react-native-worklets';
 import {
   createShareable,
   createSynchronizable,
@@ -19,7 +22,7 @@ import {
 function mutableGuestDecorator<TValue>(
   initial: TValue,
   mutable: ShareableGuest<TValue> & Mutable<TValue>,
-  dirtyFlag: Synchronizable<boolean>
+  dirtyFlag: FixedSynchronizable<boolean>
 ): ShareableGuest<TValue> & Mutable<TValue> {
   'worklet';
   let latest = initial;
@@ -31,10 +34,10 @@ function mutableGuestDecorator<TValue>(
           latest = mutable.getSync();
         } else {
           checkInvalidReadDuringRender();
-          if (dirtyFlag.getBlocking()) {
+          if (dirtyFlag.getDirty()) {
             const uiValueGetter = (svArg: Mutable<TValue>) =>
               runOnUISync((sv) => {
-                sv.setDirtyFlag?.(false);
+                sv.setDirtyFlag(false);
                 return sv.value;
               }, svArg);
             latest = uiValueGetter(mutable as Mutable<TValue>);
@@ -127,6 +130,13 @@ function mutableGuestDecorator<TValue>(
       configurable: true,
     },
 
+    setDirtyFlag: {
+      value: () => undefined,
+      writable: true,
+      enumerable: true,
+      configurable: true,
+    },
+
     _isReanimatedSharedValue: {
       value: true,
       writable: true,
@@ -139,7 +149,7 @@ function mutableGuestDecorator<TValue>(
 }
 
 export function makeMutable<TValue>(initial: TValue): Mutable<TValue> {
-  const dirtyFlag = createSynchronizable(false);
+  const dirtyFlag = createSynchronizable(false, { fixedType: true });
 
   const shareable = createShareable<TValue, Mutable<TValue>, Mutable<TValue>>(
     UIRuntimeId,