+++ packages/react-native-reanimated/src/mutables.ts
@@ -69,6 +69,8 @@ export function makeMutable<TValue>(initial: TValue): Mutable<TValue> {
       listeners.delete(id);
     },
 
+    setDirtyFlag: () => undefined,
+
     _isReanimatedSharedValue: true,
   };